Move kdf to an internal folder
